pyproject.toml
[project]
name = "cube-mcp-server"
version = "0.1.0"
description = "MCP_Server to connect to a Cube Data Store"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "mcp[cli]>=1.8.1",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
